+++ src/main/java/ua/bychkovskyy/math/impl/SubtractionOperation.java
@@ -7,7 +7,7 @@ public class SubtractionOperation implements Operation {
     private final Logger logger = Logger.getLogger(getClass());
 
     @Override
-    public double doOperation(long a, long b) {
+    public double doOperation(double a, double b) {
         logger.info(String.format("Subtraction operands '%s' and '%s'", a, b));
         return a - b;
     }
